feat(deserialize): move html deserializer to its own folder
