+++ packages/slate-plugins/src/deserializers/index.ts
@@ -1,0 +1,1 @@
+export * from './htmlDeserialize';
