+++ packages/slate-plugins/src/elements/blockquote/deserializeBlockquote.ts
@@ -1,4 +1,4 @@
-import { DeserializeHtml } from 'paste-html/types';
+import { DeserializeHtml } from 'deserializers/types';
 import { BLOCKQUOTE } from './types';
 
 export const deserializeBlockquote = (): DeserializeHtml => ({
