+++ packages/slate-plugins/src/elements/table/deserializeTable.ts
@@ -1,4 +1,4 @@
-import { DeserializeHtml } from 'paste-html/types';
+import { DeserializeHtml } from 'deserializers/types';
 import { TableType } from './types';
 
 export const deserializeTable = (): DeserializeHtml => ({
