+++ packages/slate-plugins/src/marks/inline-code/deserializeInlineCode.ts
@@ -1,4 +1,4 @@
-import { DeserializeHtml } from 'paste-html/types';
+import { DeserializeHtml } from 'deserializers/types';
 import { MARK_CODE } from './types';
 
 export const deserializeInlineCode = (): DeserializeHtml => ({
