+++ packages/slate-plugins/src/paste-html/index.ts
@@ -1,2 +1,1 @@
-export * from './deserialize';
 export * from './withPasteHtml';
